mlops-cdk-github-action/seed_code/deploy_app/requirements.txt
aws-cdk-lib
boto3
constructs
yamldataclassconfig
python-dotenv
